Move BaseFormulaFragment into android package.
